implement reviews (add + show) for each product

+++ app/src/main/java/com/example/luqtaecommerce/domain/model/product/Rating.kt
@@ -1,6 +1,6 @@
 package com.example.luqtaecommerce.domain.model.product
 
 data class Rating(
-    val average: Int,
+    val average: Double,
     val count: Int
 )
